repositories: Add chaoslab overlay
Actually, this overlay is the new home of the (already registered)
"frabjous" overlay. I decided to start fresh at gitlab.com with a
new name.

As soon	this PR	is accepted, I'll push a news item in the old
overlay, notifying potential users about the changes and requesting
them to migrate to the new overlay.

The old (frabjous) overlay will	be removed in 30 days.