Remove gnuradio overlay
I added this overlay years ago and have not been maintaining it. It is
very unlikey to be of use to anyone except as a template for making a
more up-to-date version, and I feel it should be removed entirely. In
the unlikely event I return to maintaing it, I'll open another PR to add
it back.

Closes: https://bugs.gentoo.org/864651
Signed-off-by: hololeap <hololeap@protonmail.com>
Closes: https://github.com/gentoo/api-gentoo-org/pull/637
Signed-off-by: Florian Schmaus <flow@gentoo.org>