repositories: add another-brave-overlay
This overlay automatically generates ebuilds for the Brave web browser
using GitHub workflows. In contrast to the brave-overlay, it closely
tracks the Google Chrome ebuilds in ::gentoo, with a scheduled workflow
to identify any divergence. It also provides ebuilds for all three
release channels.

The high frequency of updates for nightly releases makes a separate
overlay more practical than a contribution to GURU.